+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file, following t
 Note that since we don't clearly distinguish between a public and private interfaces there will be changes in non-major versions that are potentially breaking. If we make breaking changes to less used interfaces we will highlight it in here.
 
 ## [Unreleased]
+- MolViewSpec
+    - Fix adding default param values to primitives for primitives_from_uri
 
 ## [v5.2.0] - 2025-10-31
 - Handle transparency updates on ImagePass